Move oauth provider logic to core & split modules
