+++ docs/source/api/oauth.rst
@@ -4,16 +4,16 @@ OAuth
 .. todo::
     Docstrings (module, models, provider)
 
-.. automodule:: indico.modules.oauth
+.. automodule:: indico.core.oauth
 
 
 Models
 ++++++
 
-.. automodule:: indico.modules.oauth.models.applications
+.. automodule:: indico.core.oauth.models.applications
     :members:
     :undoc-members:
 
-.. automodule:: indico.modules.oauth.models.tokens
+.. automodule:: indico.core.oauth.models.tokens
     :members:
     :undoc-members:
