+++ indico/core/oauth/logger.py
@@ -1,0 +1,11 @@
+# This file is part of Indico.
+# Copyright (C) 2002 - 2021 CERN
+#
+# Indico is free software; you can redistribute it and/or
+# modify it under the terms of the MIT License; see the
+# LICENSE file for more details.
+
+from indico.core.logger import Logger
+
+
+logger = Logger.get('oauth')
