+++ indico/modules/events/registration/api.py
@@ -11,11 +11,11 @@ from sqlalchemy.orm import joinedload
 from werkzeug.exceptions import BadRequest, Forbidden
 
 from indico.core import signals
+from indico.core.oauth import require_oauth
 from indico.modules.events.controllers.base import RHProtectedEventBase
 from indico.modules.events.models.events import Event
 from indico.modules.events.registration.models.registrations import RegistrationState
 from indico.modules.events.registration.util import build_registration_api_data, build_registrations_api_data
-from indico.modules.oauth.oauth2 import require_oauth
 from indico.web.rh import RH
